Some additions and cleanup
- Combined files of DPLL variations into one file
- Added and fixed visualisation for uDPLL and DPLLple
- Enabled generation of cnfs based on config files
- Added a few new callable functions for benchmark tests
- Cleanup
- updated README
