+++ main.py
@@ -7,14 +7,11 @@ import src.alg.bruteforce as br
 import src.timing.measure as ms
 import src.timing.algComparison as comp
 import src.alg.dpll as dpll
-import src.alg.dpll_unit as udpll
-import src.alg.dpll_unit_ple as pledpll
 import src.alg.dp as dp
 import src.alg.dpll_visual as dpll_visual
-import src.alg.dpll_unit_visual as udpll_visual
 
 # functioning code! returns a serious KNF
-ARGUMENTS = {"-bf", "-dpll","-dpll_visual", "-udpll", "-udpllple", "-udpll_visual", "-dp", "-DPLLcomp", "-HEURcomp", "-generateCNF"}
+ARGUMENTS = {"-bf", "-dpll","-dpll_visual", "-udpll", "-dpllple","-dpllple_visual", "-udpllple", "-udpll_visual", "-dp", "-DPLLcomp", "-HEURcomp", "-generateCNF","-multiHEURcomp","-multiHEURcomp_dpll","-multiDPLLcomp"}
 UNDEFINED = 0
 
 specifiedArgument, path = arg.getArguments(ARGUMENTS)
@@ -23,7 +20,7 @@ specifiedArgument, path = arg.getArguments(ARGUMENTS)
 if path != UNDEFINED:
     
     # also check weather a comparison of algorithms should be done since the path specified will then be for folder not files
-    if specifiedArgument not in ["-DPLLcomp","-HEURcomp","-generateCNF"]:
+    if specifiedArgument not in ["-DPLLcomp","-HEURcomp","-generateCNF","-multiHEURcomp","-multiDPLLcomp","-multiHEURcomp_dpll"]:
         KNF, properties = sid.FileReader(path)
     
         # AND: the file was a legit dimacs file
@@ -32,8 +29,7 @@ if path != UNDEFINED:
             # Initiate DPLL if parameter -dpll was given
             if specifiedArgument == "-dpll":
                 var = list()
-
-                time, output = ms.timeInSeconds(dpll.output, KNF)
+                time, output = ms.timeInSeconds(dpll.output_dpll, KNF)
 
                 satisfiable, variableAssignment = output
 
@@ -46,33 +42,19 @@ if path != UNDEFINED:
             if specifiedArgument == "-dpll_visual":
                 var = list()
 
-                (satisfiable,variableAssignment)=dpll_visual.output(KNF,path)
+                (satisfiable,variableAssignment)=dpll_visual.output_dpll_visual(KNF,path)
 
                 if satisfiable:
                     print(satisfiable, "\nThe following variable assignment satisfies input cnf:",variableAssignment)
                 else:
                     print(satisfiable)
-
-            """""
-            ##### OOOLLLDLDDDDDd
-            # Initiate the visualised version of DPLL if parameter -dpll_visual was given
-            if specifiedArgument == "-dpll_visual":
-                var = list()
-
-                (satisfiable,variableAssignment)=dpll_visual.output(KNF)
-
-                if satisfiable:
-                    print(satisfiable, "\nThe following variable assignment satisfies input cnf:",variableAssignment)
-                else:
-                    print(satisfiable)
-            """
 
 
             # Initiate DPLL with unit resolution if parameter -udpll was given
             if specifiedArgument == "-udpll":
                 var = list()
 
-                time, output = ms.timeInSeconds(udpll.output, KNF)
+                time, output = ms.timeInSeconds(dpll.output_udpll, KNF)
 
                 satisfiable, variableAssignment = output
 
@@ -82,11 +64,24 @@ if path != UNDEFINED:
                     print(satisfiable, f"in {time:.5f} Sekunden!")
             
             
-            # Initiate DPLL with unit resolution if parameter -udpllple was given
+            # Initiate DPLL with pure literal elimination if parameter -dpllple was given
+            if specifiedArgument == "-dpllple":
+                var = list()
+
+                time, output = ms.timeInSeconds(dpll.output_dpllple, KNF)
+
+                satisfiable, variableAssignment = output
+
+                if satisfiable:
+                    print(satisfiable, f"in {time:.5f} Sekunden!","\nThe following variable assignment satisfies input cnf:",variableAssignment)
+                else:
+                    print(satisfiable, f"in {time:.5f} Sekunden!")
+            
+            # Initiate DPLL with unit resolution and pure literal elimination if parameter -udpllple was given
             if specifiedArgument == "-udpllple":
                 var = list()
 
-                time, output = ms.timeInSeconds(pledpll.output, KNF)
+                time, output = ms.timeInSeconds(dpll.output_udpllple, KNF)
 
                 satisfiable, variableAssignment = output
 
@@ -100,7 +95,18 @@ if path != UNDEFINED:
             if specifiedArgument == "-udpll_visual":
                 var = list()
 
-                (satisfiable,variableAssignment)=udpll_visual.output(KNF,path)
+                (satisfiable,variableAssignment)=dpll_visual.output_udpll_visual(KNF,path)
+
+                if satisfiable:
+                    print(satisfiable, "\nThe following variable assignment satisfies input cnf:",variableAssignment)
+                else:
+                    print(satisfiable)
+
+            # Initiate the visualised version of DPLL with pure literal elimination if parameter -udpll_visual was given
+            if specifiedArgument == "-dpllple_visual":
+                var = list()
+
+                (satisfiable,variableAssignment)=dpll_visual.output_dpllple_visual(KNF,path)
 
                 if satisfiable:
                     print(satisfiable, "\nThe following variable assignment satisfies input cnf:",variableAssignment)
@@ -136,18 +142,31 @@ if path != UNDEFINED:
 
     if specifiedArgument == "-DPLLcomp":
 
-            #comp.dpllComp(path)
+            comp.dpllComp(path)
+
+
+    if specifiedArgument == "-multiDPLLcomp":
+
             comp.mutipleDPLLComp(path)
+
 
     if specifiedArgument == "-HEURcomp":
 
-            #comp.heuristicsComp(path)
-            comp.mutipleHeuristicComp(path)
+            comp.heuristicsComp(path)
+
+
+    if specifiedArgument == "-multiHEURcomp":
+
+            comp.mutipleHeuristicComp(path,"udpll")
+
+            
+    if specifiedArgument == "-multiHEURcomp_dpll":
+            comp.mutipleHeuristicComp(path,"dpll")
+            
 
     if specifiedArgument == "-generateCNF":
 
-            #genCNF.output(path)
-            genCNF.runTests(path)
+            genCNF.multipleCNFs(path)
             
 
 # second case: nothing, because no proper arguments supplied
