+++ src/timing/measure.py
@@ -3,7 +3,6 @@ import time
 # this function measures the time a function needs to do its stuff and even returns the functions return values
 
 def timeInSeconds(func, arg):
-
     start = time.perf_counter()
     
     try:
@@ -21,7 +20,7 @@ def timeInSeconds(func, arg):
 
     return (timeSpended, returnValue)
 
-
+# for comparing heuristics we need to also include information about which heuristic we want to try
 def timeInSecondsHeuristics(func, arg, heuristics):
 
     start = time.perf_counter()
